Extract InvalidRequestBodyProblemException into InvalidRequestProblemException
